2019.5.7 - PlayerInteractManager, PlayerConnection, Misc

+++ src/main/java/org/bukkit/event/entity/EntityBreakDoorEvent.java
@@ -12,7 +12,7 @@ import org.bukkit.event.entity.EntityChangeBlockEvent;
 public class EntityBreakDoorEvent
 extends EntityChangeBlockEvent {
     public EntityBreakDoorEvent(LivingEntity entity, Block targetBlock) {
-        super(entity, targetBlock, Material.AIR, 0);
+        super(entity, targetBlock, Material.AIR, (byte) 0);
     }
 
     @Override
