+++ src/main/java/org/bukkit/material/MaterialData.java
@@ -6,18 +6,17 @@ package org.bukkit.material;
 import org.bukkit.Material;
 import org.bukkit.inventory.ItemStack;
 
-public class MaterialData
-implements Cloneable {
+public class MaterialData implements Cloneable {
     private final int type;
     private byte data = 0;
 
     @Deprecated
     public MaterialData(int type) {
-        this(type, 0);
+        this(type, (byte) 0);
     }
 
     public MaterialData(Material type) {
-        this(type, 0);
+        this(type, (byte) 0);
     }
 
     @Deprecated
